Updated readme metadata to fix bug with publishing to Microsoft Learn Code Samples browser
There's a bug with publishing where `data-api-builder` is not a valid value for `products`. This will be fixed soon with a new value, but it needs to be removed in the short term to publish.

<https://github.com/MicrosoftDocs/samples/blob/main/Azure-Samples/dab-swa-todo/.samples-status.json#L10>

A new PR will be created down the road to add this metadata value back once the bug is fixed.